requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
numpy>=1.24.0
Pillow>=10.0.0
rapidocr_onnxruntime>=1.3.0
pywin32>=306; platform_system == "Windows"
pandas>=2.0.0
openpyxl>=3.1.0